source

+++ project/register/src/lib.c
@@ -819,11 +819,11 @@ void spi_receive(uint8_t chip_select, uint8_t *data, uint32_t size)
     spi_send_receive(chip_select, 0, data, size);
 }
 
-
 //------------FUNCTION------------//
 
-float map(float x, float in_min, float in_max, float out_min, float out_max)
+double map(double x, double in_min, double in_max, double out_min, double out_max)
 {
   return (x - in_min) * (out_max - out_min) / (in_max - in_min) + out_min;
 }
+
